Fixed some bugs and tests

+++ src/main/java/au/com/numbrcrunchr/domain/LoanBalanceCalculator.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class LoanBalanceCalculator implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    public Long calculateLoanBalance(Long askingPrice, Long purchaseCosts,
-            Long deposit) {
+    public double calculateLoanBalance(double askingPrice,
+            double purchaseCosts, double deposit) {
         return askingPrice + purchaseCosts - deposit;
     }
